requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn
ultralytics
torch
torchvision
numpy
Pillow
python-multipart


